+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
-## [0.0.0]
+## [UNRELEASED]
 ### Changed
 - Update storage path according to core changes
+- Add tabs to config page
+
+### Added
+- Keep last playback to restore it even if player destroyed
+- Properly handle alarm creating and destroying player
+- Add button to play playlist firectly from app
 
 ## [1.0.0] - 2022-01-29
 ### Added